Speed up database insertion by using bulk inserts and caching some tables in memory
Defined manual mappings from tables to classes for SQLAlchemy. This allows us to use the bulk_insert_mappings method. SQLAlchemy requires all tables to have a primary key -- these tables now have a two-column PK of the foreign keys they reference. For example, the foreign key of a row in article_classes is (a_id, c_id). Adding a primary key implicitly adds a uniqueness guarantee. This change should be harmless since we should never map the same class to the same article more than once.

The classes, types, and dbpedia_classes tables are small compared to the page and redirect tables -- they should fit in memory. We can prevent a roundtrip to the database by just caching the id and type/class/dbpedia_class in a dictionary.

Commented out the step to create hypernyms -- since this step defines mappings from classes to dbpedia_classes, it can be done at the very end, after we've created all classes. This should reduce the amount of duplicate work. Still need to add it back in.