Change bulk insertion to handle invalid article_id foreign keys
It turns out that some article IDs in the data dump are not in the page table. MySQL raises an IntegrityError when trying to insert a foreign key into an article that doesn't exist. Added a try/except to catch this case. We should investigate this issue further. Are the data dump and page/redirect tables out of sync? Is the Wikipedia dump incomplete?