Move around commit() points to prevent IntegrityErrors
It turns out that not all IntegrityErrors are caused by invalid article ids. Sometimes, mysql can't find a class or type that should have already been flushed to the database. I think this is due to not issuing a timely commit(). I attempted to move around commit() statements to solve this issue, but I'm still not 100% certain that I fixed the root problem.

For future reference, if we find that some data is missing, this issue is worth investigating.