Use WikipediaBase-style Wikipedia classes
Because WikipediaBase is a long-standing production system, we should
use its convention for Wikipedia classes. The convention is to prepend
"wikipedia-" to class names, and use dashes instead of spaces.

For backwards compatibility and flexibility, I made it such that you can
specify both wikipedia classes (e.g. "wikipedia-french-commune") and
strings ("french commune") to query methods. We might want to revisit
this later on.